wrapper: Fix NativeLong conversion at lowLevelApi
When creating a long array from NativeLong array, we should force
each element to be treated as unsigned value.

+++ pkcs11wrapper/lib.pkcs11wrapper/src/main/java/ru/rutoken/pkcs11wrapper/lowlevel/jna/Pkcs11JnaLowLevelApi.java
@@ -95,7 +95,7 @@ public class Pkcs11JnaLowLevelApi implements IPkcs11LowLevelApi {
             return;
         if ((null != longs && null != nativeLongs) && longs.length == nativeLongs.length) {
             for (int l = 0; l < nativeLongs.length; l++) {
-                longs[l] = nativeLongs[l].longValue();
+                longs[l] = unsigned(nativeLongs[l]);
             }
         } else {
             throw new IllegalArgumentException("arrays are not in same state");
